image/test/mochitest/test_bug671906.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=671906
-->
<head>
  <title>Test for Bug 671906</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=671906">Mozilla Bug 671906</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

var first, second, third;
var correct, val1, val2;

SimpleTest.waitForExplicitFinish();

async function snapshotFirst()
{
  var iframeelem = document.getElementById('test-iframe');
  first = await snapshotWindow(iframeelem, false);

  iframeelem.onload = snapshotSecond;
  iframeelem.src = "http://example.com/tests/image/test/mochitest/bug671906-iframe.html";
}

async function snapshotSecond()
{
  var iframeelem = document.getElementById('test-iframe');
  second = await snapshotWindow(iframeelem, false);

  // We must have loaded the image again, because the principals for the
  // loading document are different.
  [correct, val1, val2] = compareSnapshots(first, second, false);
  ok(correct, "Image should have changed after changing the iframe's src.");

  iframeelem.onload = snapshotThird;
  iframeelem.src = "http://mochi.test:8888/tests/image/test/mochitest/bug671906-iframe.html";
}

async function snapshotThird()
{
  var iframeelem = document.getElementById('test-iframe');
  third = await snapshotWindow(iframeelem, false);

  // We must have loaded the image again, because the principals for the
  // loading document are different.
  [correct, val1, val2] = compareSnapshots(second, third, false);
  ok(correct, "Image should have changed after changing the iframe's src.");

  // We must have looped back to the first image, because the sjs only sends
  // one of two images.
  [correct, val1, val2] = compareSnapshots(first, third, true);
  ok(correct, "Image should be the same on the third load.");

  SimpleTest.finish();
}

</script>
</pre>
<div id="content"> <!-- style="display: none" -->
<iframe id="test-iframe" src="http://mochi.test:8888/tests/image/test/mochitest/bug671906-iframe.html" onload="snapshotFirst()"></iframe>
</div>
</body>
</html>
